Renamed component to be alphabetical. Refined individual control story.
